Test connection in connector node module
